Fix auto scaling issue
Windows Forms auto scaling works great when scaling up. It doesn't work
very well when scaling down. Thus, you need to design the form on the
monitor with the lowest DPI your application supports.

This changes the design to look good on a 96 DPI monitor. This should
also ensure it looks good on a high DPI display.

While at it I've also changed the font to use the operating system font
instead of the default font used by Windows Forms.